some little refactoring
