+++ bot/.env.dist
@@ -1,3 +1,4 @@
 TELEGRAM_BOT_TOKEN=
 SQLITE_FILE=name.db
 OPENAI_API_KEY=
+TIMEZONE=Europe/Moscow
