+++ bot/common/cache.py
@@ -1,3 +1,4 @@
+"""Cache for saving langchain agent for every user."""
 from cachetools import TTLCache
 
 agent_cache = TTLCache(maxsize=100, ttl=60)
