+++ bot/pyproject.toml
@@ -18,7 +18,6 @@ faster-whisper = "^0.8.0"
 openai = "^0.28.0"
 langchain-experimental = "^0.0.14"
 cachetools = "^5.3.1"
-pytz = "^2023.3.post1"
 
 
 [build-system]
