+++ bot/schemas/booking.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+
+class Reservation(BaseModel):
+    master: str | None = None
+    beauty_procedure: str
+    date: str
+    time: str
+    price: float
+    duration: int
