+++ bot/schemas/conversation_chain.py
@@ -1,0 +1,13 @@
+from pydantic import BaseModel
+
+
+class Booking(BaseModel):
+    action: str
+    master_name: str | None = None
+    beauty_service: str | None = None
+    booking_date: str
+    booking_time: str | None = None
+
+
+class ChatMessage(BaseModel):
+    text: str
