+++ bot/schemas/user.py
@@ -2,6 +2,6 @@ from pydantic import BaseModel
 
 
 class User(BaseModel):
-    user_id: int
+    id: int
     phone: str
     name: str
