--- setup.cfg
@@ -1,44 +1,0 @@
-[flake8]
-show-source = true
-statistics = false
-doctests = true
-
-# Plugins:
-max-cognitive-score = 20
-max-complexity = 7
-max-annotation-complexity = 4
-max-line-length = 135
-max-arguments = 8
-max-local-variables = 7
-max-tuple-unpack-length = 6
-max-try-body-length = 4
-max-raises = 5
-strictness = long
-docstring-style = google
-
-inline-quotes = "
-
-classmethod-decorators = classmethod
-
-exclude =
-  __pycache__
-  **/tests
-  **/__init__.py
-
-
-ignore =
-    D100
-
-
-[isort]
-# isort configuration:
-# https://github.com/timothycrosley/isort/wiki/isort-Settings
-include_trailing_comma = true
-use_parentheses = true
-
-
-[tool.black]
-line-length = 135
-target-version = ['py310']
-include = '\.pyi?$'
-
